Updated the specfile for proper systemd service handling

+++ prometheusListener.spec
@@ -13,7 +13,7 @@
 %define _name prometheusListener
 %define _prefix /opt
 %define _version 2.00.00
-%define _rel 0
+%define _rel 1
 %define _binaryname prometheusSDlistener
 
 Name:       prometheusListener
@@ -26,10 +26,7 @@ License:    GPL2.0
 URL:        https://git.famillegratton.net:3000/monitoring/prometheusListener
 
 Source0:    %{name}-%{_version}.tar.gz
-#BuildArchitectures: x86_64
-BuildRequires: gcc
-#Requires:
-#Obsoletes:
+BuildRequires: gcc systemd-rpm-macros
 
 %description
 Prometheus File-based Service Discovery listener
@@ -54,45 +51,32 @@ exit 0
 install -d %{buildroot}/opt/sbin
 install -d %{buildroot}/etc/systemd/system/
 install -Dpm 0755 %{_sourcedir}/%{_binaryname} %{buildroot}/opt/sbin/%{_binaryname}
+# Install the systemd service file
+install -Dpm 0644 %{_sourcedir}/prometheusSDlistener.service %{buildroot}/etc/systemd/system/prometheusSDlistener.service
 
 %post
 touch /etc/prometheusSDlistener.json
 chown -R prometheus:prometheus /etc/prometheusSDlistener
-
-cat << EOF > /etc/systemd/system/prometheusSDlistener.service
-[Unit]
-Description=Prometheus SD Listener Service
-After=network.target
-
-[Service]
-User=prometheus
-Group=prometheus
-Type=simple
-ExecStart=/opt/sbin/prometheusSDlistener
-Restart=on-failure
-RestartSec=10
-# The following is for future use as the daemon does not log right now
-#StandardOutput=append:/var/log/prometheusSDlistener.log
-#StandardError=append:/var/log/prometheusSDlistener.err
-
-[Install]
-WantedBy=multi-user.target
-EOF
-
-chmod 644 /etc/systemd/system/prometheusSDlistener.service
+# Reload systemd to apply the new service
 systemctl daemon-reload
+# Enable the service, but don't start it yet
+systemctl enable prometheusSDlistener
 
 %preun
-systemctl stop prometheusSDlistener
-systemctl disable prometheusSDlistener
+# Stop and disable the service if it's running
+if [ $1 -eq 0 ]; then
+    systemctl stop prometheusSDlistener
+    systemctl disable prometheusSDlistener
+fi
 
 %postun
-rm -f /etc/systemd/system/prometheusSDlistener.service
+# Only reload daemon on uninstall
 systemctl daemon-reload
 
 %files
 %defattr(-,root,root,-)
 /opt/sbin/%{_binaryname}
+/etc/systemd/system/prometheusSDlistener.service
 
 %changelog
 * Tue Sep 24 2024 RPM Builder <builder@famillegratton.net> 2.00.00-0
